+++ src/peer/tracks_request.rs
@@ -10,7 +10,7 @@ use tracerr::Traced;
 
 use crate::{
     media::{
-        AudioTrackConstraints, DeviceVideoTrackConstraints,
+        AudioSource, AudioTrackConstraints, DeviceVideoTrackConstraints,
         DisplayVideoTrackConstraints, MediaKind, MediaStreamSettings,
         TrackConstraints, VideoSource, track::local,
     },
@@ -23,9 +23,13 @@ use crate::{
 #[derive(Caused, Clone, Copy, Debug, Display, Eq, PartialEq)]
 #[cause(error = platform::Error)]
 pub enum TracksRequestError {
-    /// [`TracksRequest`] contains multiple [`AudioTrackConstraints`].
-    #[display("only one audio track is allowed in SimpleTracksRequest")]
-    TooManyAudioTracks,
+    /// [`TracksRequest`] contains multiple Device [`AudioTrackConstraints`].
+    #[display("only one device audio track is allowed in SimpleTracksRequest")]
+    TooManyDeviceAudioTracks,
+
+    /// [`TracksRequest`] contains multiple Display [`AudioTrackConstraints`].
+    #[display("only one display audio track is allowed in SimpleTracksRequest")]
+    TooManyDisplayAudioTracks,
 
     /// [`TracksRequest`] contains multiple [`DeviceVideoTrackConstraints`].
     #[display("only one device video track is allowed in SimpleTracksRequest")]
@@ -39,9 +43,13 @@ pub enum TracksRequestError {
     #[display("SimpleTracksRequest should have at least one track")]
     NoTracks,
 
-    /// Provided multiple audio [`local::Track`]s.
-    #[display("provided multiple audio MediaStreamTracks")]
-    ExpectedAudioTracks,
+    /// Provided multiple device audio [`local::Track`]s.
+    #[display("provided multiple device audio MediaStreamTracks")]
+    ExpectedDeviceAudioTracks,
+
+    /// Provided multiple display audio [`local::Track`]s.
+    #[display("provided multiple display audio MediaStreamTracks")]
+    ExpectedDisplayAudioTracks,
 
     /// Provided multiple device video [`local::Track`]s.
     #[display("provided multiple device video MediaStreamTracks")]
@@ -71,8 +79,11 @@ pub enum TracksRequestError {
 /// [3]: https://w3.org/TR/mediacapture-streams#mediastream
 #[derive(Debug, Default)]
 pub struct TracksRequest {
-    /// [`AudioTrackConstraints`] of [`local::Track`]s to be applied.
-    audio: HashMap<TrackId, AudioTrackConstraints>,
+    /// Device [`AudioTrackConstraints`] of [`local::Track`]s to be applied.
+    device_audio: HashMap<TrackId, AudioTrackConstraints>,
+
+    /// Display [`AudioTrackConstraints`] of [`local::Track`]s to be applied.
+    display_audio: HashMap<TrackId, AudioTrackConstraints>,
 
     /// [`DeviceVideoTrackConstraints`] of [`local::Track`]s to be applied.
     device_video: HashMap<TrackId, DeviceVideoTrackConstraints>,
@@ -89,9 +100,14 @@ impl TracksRequest {
         caps: T,
     ) {
         match caps.into() {
-            TrackConstraints::Audio(audio) => {
-                drop(self.audio.insert(track_id, audio));
-            }
+            TrackConstraints::Audio(audio) => match audio {
+                AudioSource::Device(device) => {
+                    drop(self.device_audio.insert(track_id, device));
+                }
+                AudioSource::Display(display) => {
+                    drop(self.display_audio.insert(track_id, display));
+                }
+            },
             TrackConstraints::Video(video) => match video {
                 VideoSource::Device(device) => {
                     drop(self.device_video.insert(track_id, device));
@@ -108,8 +124,11 @@ impl TracksRequest {
 /// and must have at least one track of any kind.
 #[derive(Debug)]
 pub struct SimpleTracksRequest {
-    /// [`AudioTrackConstraints`] of a [`local::Track`] to be applied.
-    audio: Option<(TrackId, AudioTrackConstraints)>,
+    /// Display [`AudioTrackConstraints`] of a [`local::Track`] to be applied.
+    display_audio: Option<(TrackId, AudioTrackConstraints)>,
+
+    /// Device [`AudioTrackConstraints`] of a [`local::Track`] to be applied.
+    device_audio: Option<(TrackId, AudioTrackConstraints)>,
 
     /// [`DisplayVideoTrackConstraints`] of a [`local::Track`] to be applied.
     display_video: Option<(TrackId, DisplayVideoTrackConstraints)>,
@@ -126,8 +145,10 @@ impl SimpleTracksRequest {
     ///
     /// - [`TracksRequestError::InvalidAudioTrack`] when some audio track from
     ///   the provided [`local::Track`]s not satisfies contained constrains.
-    /// - [`TracksRequestError::ExpectedAudioTracks`] when the provided
-    ///   [`HashMap`] doesn't have the expected audio track.
+    /// - [`TracksRequestError::ExpectedDeviceAudioTracks`] when the provided
+    ///   [`HashMap`] doesn't have the expected device audio track.
+    /// - [`TracksRequestError::ExpectedDisplayAudioTracks`] when the provided
+    ///   [`HashMap`] doesn't have the expected display audio track.
     /// - [`TracksRequestError::InvalidVideoTrack`] when some device video track
     ///   from the provided [`HashMap`] doesn't satisfy contained constrains.
     /// - [`TracksRequestError::ExpectedDeviceVideoTracks`] when the provided
@@ -148,12 +169,18 @@ impl SimpleTracksRequest {
 
         let mut display_video_tracks = Vec::new();
         let mut device_video_tracks = Vec::new();
-        let mut audio_tracks = Vec::new();
+        let mut device_audio_tracks = Vec::new();
+        let mut display_audio_tracks = Vec::new();
         for track in tracks {
             match track.kind() {
-                MediaKind::Audio => {
-                    audio_tracks.push(track);
-                }
+                MediaKind::Audio => match track.media_source_kind() {
+                    MediaSourceKind::Device => {
+                        device_audio_tracks.push(track);
+                    }
+                    MediaSourceKind::Display => {
+                        display_audio_tracks.push(track);
+                    }
+                },
                 MediaKind::Video => match track.media_source_kind() {
                     MediaSourceKind::Device => {
                         device_video_tracks.push(track);
@@ -165,15 +192,26 @@ impl SimpleTracksRequest {
             }
         }
 
-        if let Some((id, audio)) = &self.audio {
-            if let Some(track) = audio_tracks.into_iter().next() {
-                if audio.satisfies(track.as_ref()).await {
+        if let Some((id, device_audio)) = &self.device_audio {
+            if let Some(track) = device_audio_tracks.into_iter().next() {
+                if device_audio.satisfies(track.as_ref()).await {
                     drop(parsed_tracks.insert(*id, track));
                 } else {
                     return Err(tracerr::new!(InvalidAudioTrack));
                 }
             }
         }
+
+        if let Some((id, display_audio)) = &self.display_audio {
+            if let Some(track) = display_audio_tracks.into_iter().next() {
+                if display_audio.satisfies(track.as_ref()).await {
+                    drop(parsed_tracks.insert(*id, track));
+                } else {
+                    return Err(tracerr::new!(InvalidAudioTrack));
+                }
+            }
+        }
+
         if let Some((id, device_video)) = &self.device_video {
             if let Some(track) = device_video_tracks.into_iter().next() {
                 if device_video.satisfies(track.as_ref()).await {
@@ -183,6 +221,7 @@ impl SimpleTracksRequest {
                 }
             }
         }
+
         if let Some((id, display_video)) = &self.display_video {
             if let Some(track) = display_video_tracks.into_iter().next() {
                 if display_video.satisfies(track.as_ref()).await {
@@ -204,9 +243,13 @@ impl SimpleTracksRequest {
     ///
     /// # Errors
     ///
-    /// - [`TracksRequestError::ExpectedAudioTracks`] when
-    ///   [`SimpleTracksRequest`] contains [`AudioTrackConstraints`], but the
-    ///   provided [`MediaStreamSettings`] doesn't and these
+    /// - [`TracksRequestError::ExpectedDeviceAudioTracks`] when
+    ///   [`SimpleTracksRequest`] contains Device [`AudioTrackConstraints`],
+    ///   but the provided [`MediaStreamSettings`] doesn't and these Device
+    ///   [`AudioTrackConstraints`] are important.
+    /// - [`TracksRequestError::ExpectedDisplayAudioTracks`] when
+    ///   [`SimpleTracksRequest`] contains Display [`AudioTrackConstraints`],
+    ///   but the provided [`MediaStreamSettings`] doesn't and these Display
     ///   [`AudioTrackConstraints`] are important.
     /// - [`TracksRequestError::ExpectedDeviceVideoTracks`] when
     ///   [`SimpleTracksRequest`] contains [`DeviceVideoTrackConstraints`], but
@@ -222,14 +265,24 @@ impl SimpleTracksRequest {
     ) -> Result<(), Traced<TracksRequestError>> {
         let other = other.into();
 
-        if let Some((_, audio_caps)) = &self.audio {
-            if !other.is_audio_enabled() {
-                if audio_caps.required() {
+        if let Some((_, device_audio_caps)) = &self.device_audio {
+            if !other.is_device_audio_enabled() {
+                if device_audio_caps.required() {
                     return Err(tracerr::new!(
-                        TracksRequestError::ExpectedAudioTracks
+                        TracksRequestError::ExpectedDeviceAudioTracks
                     ));
                 }
-                drop(self.audio.take());
+                drop(self.device_audio.take());
+            }
+        }
+        if let Some((_, display_audio_caps)) = &self.display_audio {
+            if !other.is_display_audio_enabled() {
+                if display_audio_caps.required() {
+                    return Err(tracerr::new!(
+                        TracksRequestError::ExpectedDisplayAudioTracks
+                    ));
+                }
+                drop(self.display_audio.take());
             }
         }
         if let Some((_, device_video_caps)) = &self.device_video {
@@ -253,9 +306,14 @@ impl SimpleTracksRequest {
             }
         }
 
-        if other.is_audio_enabled() {
-            if let Some((_, audio)) = self.audio.as_mut() {
-                audio.merge(other.get_audio().clone());
+        if other.is_device_audio_enabled() {
+            if let Some((_, device_audio)) = self.device_audio.as_mut() {
+                device_audio.merge(other.get_device_audio().clone());
+            }
+        }
+        if other.is_display_audio_enabled() {
+            if let Some((_, display_audio)) = self.display_audio.as_mut() {
+                display_audio.merge(other.get_display_audio().clone());
             }
         }
         if other.is_display_video_enabled() {
@@ -282,8 +340,8 @@ impl TryFrom<TracksRequest> for SimpleTracksRequest {
 
     fn try_from(value: TracksRequest) -> Result<Self, Self::Error> {
         use TracksRequestError::{
-            NoTracks, TooManyAudioTracks, TooManyDeviceVideoTracks,
-            TooManyDisplayVideoTracks,
+            NoTracks, TooManyDeviceAudioTracks, TooManyDeviceVideoTracks,
+            TooManyDisplayAudioTracks, TooManyDisplayVideoTracks,
         };
 
         #[expect(clippy::else_if_without_else, reason = "more readable")]
@@ -291,20 +349,31 @@ impl TryFrom<TracksRequest> for SimpleTracksRequest {
             return Err(TooManyDeviceVideoTracks);
         } else if value.display_video.len() > 1 {
             return Err(TooManyDisplayVideoTracks);
-        } else if value.audio.len() > 1 {
-            return Err(TooManyAudioTracks);
+        } else if value.device_audio.len() > 1 {
+            return Err(TooManyDeviceAudioTracks);
+        } else if value.display_audio.len() > 1 {
+            return Err(TooManyDisplayAudioTracks);
         } else if value.device_video.is_empty()
             && value.display_video.is_empty()
-            && value.audio.is_empty()
+            && value.device_audio.is_empty()
+            && value.display_video.is_empty()
         {
             return Err(NoTracks);
         }
 
-        let mut req =
-            Self { audio: None, device_video: None, display_video: None };
+        let mut req = Self {
+            device_audio: None,
+            display_audio: None,
+            device_video: None,
+            display_video: None,
+        };
         #[expect(clippy::iter_over_hash_type, reason = "order doesn't matter")]
-        for (id, audio) in value.audio {
-            drop(req.audio.replace((id, audio)));
+        for (id, device) in value.device_audio {
+            drop(req.device_audio.replace((id, device)));
+        }
+        #[expect(clippy::iter_over_hash_type, reason = "order doesn't matter")]
+        for (id, display) in value.display_audio {
+            drop(req.display_audio.replace((id, display)));
         }
         #[expect(clippy::iter_over_hash_type, reason = "order doesn't matter")]
         for (id, device) in value.device_video {
@@ -323,12 +392,18 @@ impl From<&SimpleTracksRequest> for MediaStreamSettings {
     fn from(request: &SimpleTracksRequest) -> Self {
         let mut constraints = Self::new();
 
-        if let Some((_, audio)) = &request.audio {
-            constraints.audio(audio.clone());
+        if let Some((_, device_audio)) = &request.device_audio {
+            constraints.device_audio(device_audio.clone());
         }
+
+        if let Some((_, display_audio)) = &request.display_audio {
+            constraints.display_audio(display_audio.clone());
+        }
+
         if let Some((_, device_video)) = &request.device_video {
             constraints.device_video(device_video.clone());
         }
+
         if let Some((_, display_video)) = &request.display_video {
             constraints.display_video(display_video.clone());
         }