+++ tests/room/room.rs
@@ -2493,7 +2493,7 @@ mod set_local_media_settings {
         let (audio_track, video_track) = get_test_required_tracks();
 
         let mut constraints = api::MediaStreamSettings::new();
-        constraints.audio(api::AudioTrackConstraints::new());
+        constraints.device_audio(api::AudioTrackConstraints::new());
 
         JsFuture::from(room_handle.set_local_media_settings(
             &constraints,
@@ -2543,7 +2543,7 @@ mod set_local_media_settings {
                 .await;
 
         let mut constraints = api::MediaStreamSettings::new();
-        constraints.audio(api::AudioTrackConstraints::new());
+        constraints.device_audio(api::AudioTrackConstraints::new());
 
         let room_handle = api::RoomHandle::from(room.new_handle());
         room_handle.on_failed_local_media(cb.into()).unwrap();
@@ -2595,7 +2595,7 @@ mod set_local_media_settings {
 
             let mut constraints = api::MediaStreamSettings::new();
             if add_audio {
-                constraints.audio(api::AudioTrackConstraints::new());
+                constraints.device_audio(api::AudioTrackConstraints::new());
             }
             if add_video {
                 constraints